Agregando componente para manejar errores. Cambios de manejo de estado de modales. Cambio de manejo de rutas en componente App
