+++ my-app/src/components/modal/index.jsx
@@ -1,15 +1,16 @@
 import React, { Component } from 'react';
 import "./style.css";
 import { Link } from "react-router-dom";
-import PasswordInput from '../password/index.jsx';
+import PasswordInput from '../password/index';
 import EntryButton from '../Buttons/EntryButton';
-import FlatButton from '../Buttons/flatButton/index.jsx';
+import FlatButton from '../Buttons/flatButton/index';
+import TextErrors from '../textErrors/index';
 import { db, firebaseApp } from '../../firebase/index';
 
 class Modal extends Component {
   constructor(props){
   super(props)
-  this.state = {pin: ""}
+  this.state = {pin: "",textError:""}
   this.handlePinChange = this.handlePinChange.bind(this);
   this.handlePinFromFirebase = this.handlePinFromFirebase.bind(this);
   }
@@ -24,19 +25,23 @@ class Modal extends Component {
       const pinFirebase = doc.data().pin
       return pinFirebase
   }).then(pinFirebase => {
-      if(pinFirebase == this.state.pin){
-        
-       console.log('son iguales')
+      if(pinFirebase === this.state.pin){
+        const {close} = this.props;
+        close();
       }else{
-        console.error('No son iguales')
+        const {open}= this.props;
+        open()
       }
+  }).catch(()=>{
+    console.error('Clave es incorrecta');
+    this.setState({textError: "Tu clave es incorrecta"})
   })
   ;
 
   }
 
 render(){
-  const {pin} = this.state;
+  const {pin, textError} = this.state;
   const {open, close} = this.props;
   const classes = `overlay ${open ? "open" : "closed"}`;
   console.log(pin);
@@ -48,6 +53,7 @@ render(){
         </div>
         <div className="modal-input">
           <PasswordInput labelText="Ingresa tu clave de seguridad" value={pin} onChange={this.handlePinChange}/>
+          <TextErrors text={textError}/>
         </div>
         <div className="modal-footer">
           <Link to="/">
