+++ my-app/src/components/modal/style.css
@@ -12,7 +12,7 @@
 
 .modal {
     width: 500px;
-    height: 250px;
+    height: 300px;
     border-radius: 5px;
     background-color: whitesmoke;
     padding: 0.5rem;
