+++ my-app/src/components/textErrors/index.jsx
@@ -1,0 +1,10 @@
+import React, {Component} from 'react';
+import './style.css';
+
+const TextErrors = ({text}) => {
+    return(
+        <p className="text-errors">{text}</p>
+    )
+}
+
+export default TextErrors;
