+++ my-app/src/components/textErrors/style.css
@@ -1,0 +1,5 @@
+.text-errors {
+    color: red;
+    font-size: small;
+    text-align: center;
+}
